+++ rdkit/Chem/GraphDescriptors.py
@@ -29,10 +29,6 @@ periodicTable = rdchem.GetPeriodicTable()
 _log2val = math.log(2)
 
 
-def _log2(x):
-  return math.log(x) / _log2val
-
-
 def _VertexDegrees(mat, onlyOnes=0):
   """  *Internal Use Only*
 
@@ -89,10 +85,10 @@ def _pyHallKierAlpha(m):
     else:
       rA = PeriodicTable.nameTable[symb][5]
       alpha = rA / rC - 1
-    print(atom.GetIdx(), atom.GetSymbol(), alpha)
+    # print(atom.GetIdx(), atom.GetSymbol(), alpha)
     alphaSum += alpha
   return alphaSum
-#HallKierAlpha.version="1.0.2"
+# HallKierAlpha.version="1.0.2"
 
 
 def Ipc(mol, avg=0, dMat=None, forceDMat=0):
@@ -130,57 +126,57 @@ def _pyKappa1(mol):
   """ Hall-Kier Kappa1 value
 
    From equations (58) and (59) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   P1 = mol.GetNumBonds(1)
   A = mol.GetNumHeavyAtoms()
   alpha = HallKierAlpha(mol)
   denom = P1 + alpha
   if denom:
-    kappa = (A + alpha) * (A + alpha - 1)**2 / denom**2
+    kappa = (A + alpha) * (A + alpha - 1) ** 2 / denom ** 2
   else:
     kappa = 0.0
   return kappa
-#Kappa1.version="1.0.0"
+# Kappa1.version="1.0.0"
 
 
 def _pyKappa2(mol):
   """  Hall-Kier Kappa2 value
 
    From equations (58) and (60) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   P2 = len(Chem.FindAllPathsOfLengthN(mol, 2))
   A = mol.GetNumHeavyAtoms()
   alpha = HallKierAlpha(mol)
-  denom = (P2 + alpha)**2
+  denom = (P2 + alpha) ** 2
   if denom:
-    kappa = (A + alpha - 1) * (A + alpha - 2)**2 / denom
+    kappa = (A + alpha - 1) * (A + alpha - 2) ** 2 / denom
   else:
     kappa = 0
   return kappa
-#Kappa2.version="1.0.0"
+# Kappa2.version="1.0.0"
 
 
 def _pyKappa3(mol):
   """  Hall-Kier Kappa3 value
 
    From equations (58), (61) and (62) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   P3 = len(Chem.FindAllPathsOfLengthN(mol, 3))
   A = mol.GetNumHeavyAtoms()
   alpha = HallKierAlpha(mol)
-  denom = (P3 + alpha)**2
+  denom = (P3 + alpha) ** 2
   if denom:
     if A % 2 == 1:
-      kappa = (A + alpha - 1) * (A + alpha - 3)**2 / denom
+      kappa = (A + alpha - 1) * (A + alpha - 3) ** 2 / denom
     else:
-      kappa = (A + alpha - 2) * (A + alpha - 3)**2 / denom
+      kappa = (A + alpha - 2) * (A + alpha - 3) ** 2 / denom
   else:
     kappa = 0
   return kappa
-#Kappa3.version="1.0.0"
+# Kappa3.version="1.0.0"
 
 HallKierAlpha = lambda x: rdMolDescriptors.CalcHallKierAlpha(x)
 HallKierAlpha.version = rdMolDescriptors._CalcHallKierAlpha_version
@@ -194,7 +190,7 @@ Kappa3.version = rdMolDescriptors._CalcKappa3_version
 
 def Chi0(mol):
   """ From equations (1),(9) and (10) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   deltas = [x.GetDegree() for x in mol.GetAtoms()]
   while 0 in deltas:
@@ -209,7 +205,7 @@ Chi0.version = "1.0.0"
 
 def Chi1(mol):
   """ From equations (1),(11) and (12) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   c1s = [x.GetBeginAtom().GetDegree() * x.GetEndAtom().GetDegree() for x in mol.GetBonds()]
   while 0 in c1s:
@@ -253,7 +249,7 @@ def _hkDeltas(mol, skipHs=1):
 
 def _pyChi0v(mol):
   """  From equations (5),(9) and (10) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   deltas = _hkDeltas(mol)
   while 0 in deltas:
@@ -265,7 +261,7 @@ def _pyChi0v(mol):
 
 def _pyChi1v(mol):
   """  From equations (5),(11) and (12) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   deltas = numpy.array(_hkDeltas(mol, skipHs=0))
   res = 0.0
@@ -278,13 +274,13 @@ def _pyChi1v(mol):
 
 def _pyChiNv_(mol, order=2):
   """  From equations (5),(15) and (16) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   **NOTE**: because the current path finding code does, by design,
   detect rings as paths (e.g. in C1CC1 there is *1* atom path of
   length 3), values of ChiNv with N >= 3 may give results that differ
   from those provided by the old code in molecules that have rings of
   size 3.
-  
+
   """
   deltas = numpy.array(
     [(1. / numpy.sqrt(hkd) if hkd != 0.0 else 0.0) for hkd in _hkDeltas(mol, skipHs=0)])
@@ -296,26 +292,26 @@ def _pyChiNv_(mol, order=2):
 
 def _pyChi2v(mol):
   """ From equations (5),(15) and (16) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   return _pyChiNv_(mol, 2)
 
 
 def _pyChi3v(mol):
   """ From equations (5),(15) and (16) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   """
   return _pyChiNv_(mol, 3)
 
 
 def _pyChi4v(mol):
   """ From equations (5),(15) and (16) of Rev. Comp. Chem. vol 2, 367-422, (1991)
-   
+
   **NOTE**: because the current path finding code does, by design,
   detect rings as paths (e.g. in C1CC1 there is *1* atom path of
   length 3), values of Chi4v may give results that differ from those
   provided by the old code in molecules that have 3 rings.
-  
+
   """
   return _pyChiNv_(mol, 4)
 
@@ -323,7 +319,7 @@ def _pyChi4v(mol):
 def _pyChi0n(mol):
   """  Similar to Hall Kier Chi0v, but uses nVal instead of valence
   This makes a big difference after we get out of the first row.
-   
+
   """
   deltas = [_nVal(x) for x in mol.GetAtoms()]
   while deltas.count(0):
@@ -335,7 +331,7 @@ def _pyChi0n(mol):
 
 def _pyChi1n(mol):
   """  Similar to Hall Kier Chi1v, but uses nVal instead of valence
-   
+
   """
   delts = numpy.array([_nVal(x) for x in mol.GetAtoms()], 'd')
   res = 0.0
@@ -349,13 +345,13 @@ def _pyChi1n(mol):
 def _pyChiNn_(mol, order=2):
   """  Similar to Hall Kier ChiNv, but uses nVal instead of valence
   This makes a big difference after we get out of the first row.
-   
+
   **NOTE**: because the current path finding code does, by design,
   detect rings as paths (e.g. in C1CC1 there is *1* atom path of
   length 3), values of ChiNn with N >= 3 may give results that differ
   from those provided by the old code in molecules that have rings of
   size 3.
-  
+
   """
   nval = [_nVal(x) for x in mol.GetAtoms()]
   deltas = numpy.array([(1. / numpy.sqrt(x) if x else 0.0) for x in nval])
@@ -368,7 +364,7 @@ def _pyChiNn_(mol, order=2):
 def _pyChi2n(mol):
   """  Similar to Hall Kier Chi2v, but uses nVal instead of valence
   This makes a big difference after we get out of the first row.
-   
+
   """
   return _pyChiNn_(mol, 2)
 
@@ -376,7 +372,7 @@ def _pyChi2n(mol):
 def _pyChi3n(mol):
   """  Similar to Hall Kier Chi3v, but uses nVal instead of valence
   This makes a big difference after we get out of the first row.
-   
+
   """
   return _pyChiNn_(mol, 3)
 
@@ -384,13 +380,13 @@ def _pyChi3n(mol):
 def _pyChi4n(mol):
   """  Similar to Hall Kier Chi4v, but uses nVal instead of valence
   This makes a big difference after we get out of the first row.
-   
+
 
   **NOTE**: because the current path finding code does, by design,
   detect rings as paths (e.g. in C1CC1 there is *1* atom path of
   length 3), values of Chi4n may give results that differ from those
   provided by the old code in molecules that have 3 rings.
-  
+
   """
   return _pyChiNn_(mol, 4)
 
@@ -474,16 +470,16 @@ def BalabanJ(mol, dMat=None, forceDMat=0):
   n = mol.GetNumAtoms()
   mu = q - n + 1
 
-  sum = 0.
+  sum_ = 0.
   nS = len(s)
   for i in range(nS):
     si = s[i]
     for j in range(i, nS):
       if adjMat[i, j] == 1:
-        sum += 1. / numpy.sqrt(si * s[j])
+        sum_ += 1. / numpy.sqrt(si * s[j])
 
   if mu + 1 != 0:
-    J = float(q) / float(mu + 1) * sum
+    J = float(q) / float(mu + 1) * sum_
   else:
     J = 0
 
@@ -492,7 +488,7 @@ def BalabanJ(mol, dMat=None, forceDMat=0):
 
 BalabanJ.version = "1.0.0"
 
-#------------------------------------------------------------------------
+# ------------------------------------------------------------------------
 #
 # Start block of BertzCT stuff.
 #
@@ -504,13 +500,12 @@ def _AssignSymmetryClasses(mol, vdList, bdMat, forceBDMat, numAtoms, cutoff):
 
      vdList: the number of neighbors each atom has
      bdMat: "balaban" distance matrix
-     
+
   """
   if forceBDMat:
     bdMat = Chem.GetDistanceMatrix(mol, useBO=1, useAtomWts=0, force=1, prefix="Balaban")
     mol._balabanMat = bdMat
 
-  atomIdx = 0
   keysSeen = []
   symList = [0] * numAtoms
   for i in range(numAtoms):
@@ -539,7 +534,7 @@ def _LookUpBondOrder(atom1Id, atom2Id, bondDic):
     tmp = 1.5
   else:
     tmp = float(tmp)
-  #tmp = int(tmp)
+  # tmp = int(tmp)
   return tmp
 
 
@@ -627,7 +622,7 @@ def BertzCT(mol, cutoff=100, dMat=None, forceDMat=1):
        Any molecule containing aromatic rings will yield different
        values with this implementation.  The new behavior is the correct
        one, so we're going to live with the breakage.
-        
+
      **NOTE** this barfs if the molecule contains a second (or
        nth) fragment that is one atom.
 
@@ -652,7 +647,7 @@ def BertzCT(mol, cutoff=100, dMat=None, forceDMat=1):
 
   bondDict, neighborList, vdList = _CreateBondDictEtc(mol, numAtoms)
   symmetryClasses = _AssignSymmetryClasses(mol, vdList, dMat, forceDMat, numAtoms, cutoff)
-  #print('Symmm Classes:',symmetryClasses)
+  # print('Symmm Classes:',symmetryClasses)
   for atomIdx in range(numAtoms):
     hingeAtomNumber = mol.GetAtomWithIdx(atomIdx).GetAtomicNum()
     atomTypeDict[hingeAtomNumber] = atomTypeDict.get(hingeAtomNumber, 0) + 1
@@ -663,7 +658,7 @@ def BertzCT(mol, cutoff=100, dMat=None, forceDMat=1):
       neighbor_iIdx = neighborList[atomIdx][i]
       NiClass = symmetryClasses[neighbor_iIdx]
       bond_i_order = _LookUpBondOrder(atomIdx, neighbor_iIdx, bondDict)
-      #print('\t',atomIdx,i,hingeAtomClass,NiClass,bond_i_order)
+      # print('\t',atomIdx,i,hingeAtomClass,NiClass,bond_i_order)
       if (bond_i_order > 1) and (neighbor_iIdx > atomIdx):
         numConnections = bond_i_order * (bond_i_order - 1) / 2
         connectionKey = (min(hingeAtomClass, NiClass), max(hingeAtomClass, NiClass))
@@ -692,4 +687,4 @@ BertzCT.version = "2.0.0"
 #
 # End block of BertzCT stuff.
 #
-#------------------------------------------------------------------------
+# ------------------------------------------------------------------------
